feat(estudio): candado en la onda, tono en tonos y metronomo con decimales, compases y mas volumen
La onda lleva un candado, puesto de entrada: mientras suena, un clic no mueve la cancion y elegir un tramo (o la banderita de un marcador) no salta a el. Rust arma el bucle cuando la cancion entra en el tramo; si va por detras, la cancion sigue y al acabarse vuelve a A en vez de pasar a la siguiente. Las teclas A y B y la lista de marcadores siguen moviendo la cancion, y en pausa la onda coloca como siempre.

El tono se cuenta en tonos (medio tono es un semitono) y se corre de cuarto, de medio o de tono entero: Rust recibe semitonos con fracciones, que rubberband y asetrate aceptan tal cual. Con un cuarto de tono se dice de que tono se parte, «G +¼».

Metronomo: ×2 y ÷2 no hacian nada con un tempo a mano, no se veian puestos (tampoco la velocidad elegida: no habia estilo para .btn.on en la barra) y la onda pintaba la rejilla del analisis tal cual. Ahora doblan tambien el tempo a mano, se ven encendidos y la onda enseña la rejilla tal como suena. El tempo se escribe con decimales, y con la cancion sonando el clic a mano entra en su siguiente pulso. Compases 2/4, 3/4, 4/4, 6/8 y sin acento.

Volumen: la cancion sube hasta el 150 % (tambien desde el reproductor) y el clic hasta el doble. Todo pasa por un limitador en la salida, que a volumen normal no toca ninguna muestra; la rueda y las flechas se paran en el 100 %.

+++ desktop/src-tauri/src/player/mod.rs
@@ -72,8 +72,9 @@ pub enum Command {
     /// Repetir de A a B (segundos de la cancion); None lo quita.
     Loop(Option<(f64, f64)>),
     Speed(f32),
-    /// El tono corrido, en semitonos (-12..12). Reabre la cancion donde iba.
-    Pitch(i32),
+    /// El tono corrido, en semitonos (-12..12, con fracciones: medio
+    /// semitono es un cuarto de tono). Reabre la cancion donde iba.
+    Pitch(f32),
     /// Ajustes del metronomo, y la rejilla de la cancion `path` si se conoce.
     Metronome {
         settings: MetronomeSettings,
